add test directory
